Fetch URL History enhancemant and a few fixes
- New Feature: Fetch URL History. Most recent AtomSpace Fetch URL is
prepopulated in the Fetch URL field. Implemented by storing in browser
local storage. Can be cleared with the Fetch popup's 'Reset' button,
or by clearing your browser cache. Fixes #103.
- Bug Fix: Capped max fetch atom count to 2500, though the practical
limit is probably below that. D3 graphs are DOM heavy and it's
currently not feasible to load up a greater number of atoms than that.
Fixes #101.
- Bug Fix: Changed short-cut for stealth version number popup from 'v'
key to Ctrl-Alt-v. Fixes #104.
- Bug Fix: Stop (hidden) simulation when click on Fetch button.